+++ services/blockchain-indexer/shared/dataService/business/index.js
@@ -48,6 +48,9 @@ const {
 	getChainAccount,
 	getMainchainID,
 	reloadBlockchainAppsStats,
+	isMainchain,
+	resolveMainchainServiceURL,
+	resolveChannelInfo,
 } = require('./interoperability');
 
 const {
@@ -99,7 +102,6 @@ const {
 	getNetworkPeersStatistics,
 } = require('./network');
 const { estimateTransactionFees } = require('./transactionsEstimateFees');
-const { isMainchain, resolveMainchainServiceURL } = require('./mainchain');
 const { invokeEndpoint } = require('./invoke');
 
 const { setFeeEstimates, getFeeEstimates } = require('./feeEstimates');
@@ -144,6 +146,9 @@ module.exports = {
 	getMainchainID,
 	getBlockchainAppsStatistics,
 	reloadBlockchainAppsStats,
+	isMainchain,
+	resolveMainchainServiceURL,
+	resolveChannelInfo,
 
 	// Token
 	tokenHasUserAccount,
@@ -193,9 +198,6 @@ module.exports = {
 	getNetworkConnectedPeers,
 	getNetworkDisconnectedPeers,
 	getNetworkPeersStatistics,
-
-	isMainchain,
-	resolveMainchainServiceURL,
 
 	invokeEndpoint,
 };
